+++ next.config.js
@@ -12,8 +12,7 @@ module.exports = {
     { dev, dir, outDir, distDir, buildId }
   ) {
     return {
-      '/': { page: '/' },
-      '/about': { page: '/about' },
+      '/': { page: '/' }
     }
   },
   images: {
