--- pages/about.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const about = () => {
-  return (
-    <div>
-      <h1>This is about us</h1>
-    </div>
-  );
-};
-
-export default about;<h1>This is about us</h1>
