+++ pages/privacy-policy/index.jsx
@@ -4,18 +4,20 @@ import Container from '../../components/common/Container';
 import PrivacyPolicy from '../../components/PrivacyPolicy/PrivacyPolicy';
 import DownlaodOurApp from '../../components/common/DownlaodOurApp';
 import SingingCompany from '../../components/Home/SingingCompany';
-import Footer from '../../components/common/Footer2';
+import Footer from '../../components/common/Footer';
 
 
 const index = () => {
   return (
     <>
       <Navbar />
-      <Container>
-        <PrivacyPolicy />
-      </Container>
-      <DownlaodOurApp />
-      <SingingCompany />
+      <main>
+        <Container>
+          <PrivacyPolicy />
+        </Container>
+        <DownlaodOurApp />
+        <SingingCompany />
+      </main>
       <Footer />
     </>
   );
